// frontends/checkpoint/src/redux/reducers/types.ts
export enum EUserAction {
    USER_LOGIN = "USER_LOGIN",
    USER_LOGIN_SUCCESS = "USER_LOGIN_SUCCESS",
    USER_LOGIN_FAIL = "USER_LOGIN_FAIL",
    USER_LOGOUT = "USER_LOGOUT",
    USER_RENEW = "USER_RENEW",
    USER_RENEW_SUCCESS = "USER_RENEW_SUCCESS",
    USER_RENEW_FAIL = "USER_RENEW_FAIL",
    CLEAR_ERROR = "CLEAR_ERROR"
}